# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenMaintainer core API",
    "version": "0.1.0",
    "description": "Core route inventory. See api.md and source schemas for complete routes and payloads."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Process health",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful read or mutation"
          },
          "202": {
            "description": "Accepted queued work, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient authority"
          },
          "404": {
            "description": "Not found within authorized scope"
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Database readiness",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful read or mutation"
          },
          "202": {
            "description": "Accepted queued work, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient authority"
          },
          "404": {
            "description": "Not found within authorized scope"
          }
        }
      }
    },
    "/webhooks/github": {
      "post": {
        "summary": "HMAC-verified webhook receipt",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful read or mutation"
          },
          "202": {
            "description": "Accepted queued work, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient authority"
          },
          "404": {
            "description": "Not found within authorized scope"
          }
        }
      }
    },
    "/api/repositories": {
      "get": {
        "summary": "Accessible repositories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful read or mutation"
          },
          "202": {
            "description": "Accepted queued work, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient authority"
          },
          "404": {
            "description": "Not found within authorized scope"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/repositories/{owner}/{repo}": {
      "get": {
        "summary": "Repository and recent records",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful read or mutation"
          },
          "202": {
            "description": "Accepted queued work, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient authority"
          },
          "404": {
            "description": "Not found within authorized scope"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/repositories/{owner}/{repo}/config": {
      "get": {
        "summary": "Maintainer-approved policy",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful read or mutation"
          },
          "202": {
            "description": "Accepted queued work, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient authority"
          },
          "404": {
            "description": "Not found within authorized scope"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ]
      },
      "put": {
        "summary": "Maintainer-approved policy",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful read or mutation"
          },
          "202": {
            "description": "Accepted queued work, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient authority"
          },
          "404": {
            "description": "Not found within authorized scope"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/repositories/{owner}/{repo}/analyze": {
      "post": {
        "summary": "Queue repository analysis",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful read or mutation"
          },
          "202": {
            "description": "Accepted queued work, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient authority"
          },
          "404": {
            "description": "Not found within authorized scope"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/repositories/{owner}/{repo}/analyses": {
      "get": {
        "summary": "Stored analysis history",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful read or mutation"
          },
          "202": {
            "description": "Accepted queued work, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient authority"
          },
          "404": {
            "description": "Not found within authorized scope"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/repositories/{owner}/{repo}/ai": {
      "post": {
        "summary": "Queue optional advisory AI",
        "parameters": [
          {
            "name": "owner",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful read or mutation"
          },
          "202": {
            "description": "Accepted queued work, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient authority"
          },
          "404": {
            "description": "Not found within authorized scope"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/jobs": {
      "get": {
        "summary": "Scoped jobs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful read or mutation"
          },
          "202": {
            "description": "Accepted queued work, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient authority"
          },
          "404": {
            "description": "Not found within authorized scope"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/jobs/{id}": {
      "get": {
        "summary": "Job status and history",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful read or mutation"
          },
          "202": {
            "description": "Accepted queued work, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient authority"
          },
          "404": {
            "description": "Not found within authorized scope"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/jobs/{id}/cancel": {
      "post": {
        "summary": "Cancel queued job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful read or mutation"
          },
          "202": {
            "description": "Accepted queued work, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient authority"
          },
          "404": {
            "description": "Not found within authorized scope"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/analytics": {
      "get": {
        "summary": "Scoped job counts and AI usage",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful read or mutation"
          },
          "202": {
            "description": "Accepted queued work, where applicable"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient authority"
          },
          "404": {
            "description": "Not found within authorized scope"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "GitHub user access token"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "om_session"
      }
    }
  }
}